Specialise layouts for performance.
